+++ torchmetrics/__init__.py
@@ -50,7 +50,9 @@ from torchmetrics.regression import (  # noqa: F401 E402
     MeanAbsoluteError,
     MeanSquaredError,
     MeanSquaredLogError,
+    PearsonCorrcoef,
     R2Score,
+    SpearmanCorrcoef,
 )
 from torchmetrics.retrieval import (  # noqa: F401 E402
     RetrievalFallOut,